# services/prediction-service/openapi.yaml
openapi: 3.0.3
info:
  title: prediction-service API
  version: "1.0.0"
  description: >
    Combina el riesgo determinístico de analytics-service con el contexto académico
    del estudiante y genera, vía OpenAI, un resumen + plan de recomendaciones.
servers:
  - url: /api/v1/prediction
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RiskSnapshot:
      type: object
      properties:
        riskLevel: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        globalAverage: { type: number }
        complianceIndex: { type: number }
        attendance: { type: number }
        failedEvaluations: { type: integer }
        trendSlope: { type: number }
    RecommendationItem:
      type: object
      properties:
        type:
          type: string
          enum:
            [
              STUDY_HABIT,
              TUTORING,
              TIME_MANAGEMENT,
              ATTENDANCE,
              SUBJECT_FOCUS,
              WELLBEING,
            ]
        title: { type: string }
        reason: { type: string }
        priority: { type: string, enum: [LOW, MEDIUM, HIGH] }
    PredictionResult:
      type: object
      properties:
        studentId: { type: string }
        periodId: { type: string }
        risk: { $ref: "#/components/schemas/RiskSnapshot" }
        summary: { type: string }
        recommendations:
          type: array
          items: { $ref: "#/components/schemas/RecommendationItem" }
        modelVersion: { type: string }
        generatedAt: { type: string, format: date-time }
security:
  - JWT: []
paths:
  /students/{studentId}/periods/{periodId}:
    get:
      summary: Generate prediction + AI recommendations for a student
      parameters:
        - in: path
          name: studentId
          required: true
          schema: { type: string }
        - in: path
          name: periodId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionResult"
  /students/{studentId}/history:
    get:
      summary: List past predictions for a student
      parameters:
        - in: path
          name: studentId
          required: true
          schema: { type: string }
        - in: query
          name: limit
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Prediction history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PredictionResult"